+++ source/Core/Autoload/UnifiedNameSpaceClassMap.php
@@ -3381,4 +3381,22 @@ return [
         'isInterface'      => false,
         'isDeprecated'     => false
     ],
+    'OxidEsales\Eshop\Core\Module\ModuleSmartyPluginDirectories'                 => [
+        'editionClassName' => \OxidEsales\EshopCommunity\Core\Module\ModuleSmartyPluginDirectories::class,
+        'isAbstract'       => false,
+        'isInterface'      => false,
+        'isDeprecated'     => true
+    ],
+    'OxidEsales\Eshop\Core\Module\ModuleSmartyPluginDirectoryRepository'         => [
+        'editionClassName' => \OxidEsales\EshopCommunity\Core\Module\ModuleSmartyPluginDirectoryRepository::class,
+        'isAbstract'       => false,
+        'isInterface'      => false,
+        'isDeprecated'     => true
+    ],
+    'OxidEsales\Eshop\Core\Module\ModuleSmartyPluginDirectoryValidator'          => [
+        'editionClassName' => \OxidEsales\EshopCommunity\Core\Module\ModuleSmartyPluginDirectoryValidator::class,
+        'isAbstract'       => false,
+        'isInterface'      => false,
+        'isDeprecated'     => true
+    ],
 ];